01_MusicStamp のiOS7対応を完了。

+++ iOS/01_MusicStamp/obj-c/TouchStamp/touchPointsViewController.m
@@ -17,7 +17,7 @@
 @interface touchPointsViewController () {    
     absPatternRecognizer *recognizer_;
     SystemSoundID sid_;
-    int prev_touches_;
+    NSUInteger prev_touches_;
 }
 
 -(int)getUnitLength:(double)lengthInMiliMeter;
